src/Web/Controllers/ReportController.cs
using System.Web.Mvc;
using Dimensional.TinyReturns.Core;
using Dimensional.TinyReturns.Core.PortfolioReportingContext.Services.PublicWebReport;

namespace Dimensional.TinyReturns.Web.Controllers
{
    public class ReportController : Controller
    {
        private readonly PublicWebReportFacade _publicWebReportFacade;

        public ReportController()
        {
            _publicWebReportFacade = MasterFactory.GetPublicWebReportFacade();
        }

        // Used for tests
        public ReportController(
            PublicWebReportFacade publicWebReportFacade)
        {
            _publicWebReportFacade = publicWebReportFacade;
        }

        public ActionResult Index()
        {
            return View(_publicWebReportFacade.GetPortfolioPerforance());
        }
    }
}